feat: add App::new()
